added pages folder with home page
